moving `train_cl_model` and `evaluate_cl_model` to a separate file
